improve sequence of commands to deal with errors

+++ crypto.mk
@@ -1,6 +1,7 @@
 NOW := $(date +%Y-%m-%d_%H-%M)
 notes.org.gpg: notes.org
-	-rm -i $@
-	-gpg -c $<
-	-cp -v  $< /tmp/$(NOW)-$<
-	-git commit -m "notes up to $(NOW)" $@
+	-if [ -f $@ ]; then rm -i $@; fi
+	-gpg -c $< \
+	&& cp -v  $< /tmp/$(NOW)-$< \
+	&& git commit -m "notes up to $(NOW)" $@ \
+	&& git push
